+++ src/openMVG/dataio/CMakeLists.txt
@@ -21,9 +21,6 @@ target_link_libraries( openMVG_dataio openMVG_sfm ${Boost_LIBRARIES})
 if(OpenMVG_USE_OPENCV)
   target_link_libraries( openMVG_dataio ${OpenCV_LIBS})
 endif(OpenMVG_USE_OPENCV)
-if(OpenMVG_USE_ALEMBIC)
-  target_link_libraries( openMVG_dataio ${ABC_LIBRARIES})
-endif(OpenMVG_USE_ALEMBIC)
 install(TARGETS openMVG_dataio DESTINATION lib/ EXPORT openMVG-targets)
 
 #UNIT_TEST(openMVG dataio "openMVG_dataio")
